contrib: create a new internal httptrace package
